更改 DrawIndexed 方法的 vertexOffset 参数类型
将多个类和结构体中的 DrawIndexed 方法和字段的 vertexOffset 参数类型从 uint 更改为 int，包括 CommandBuffer、DXCommandBuffer、ImGuiRenderer 和 VKCommandBuffer 类，以及 IndirectDrawIndexedArgs 结构体。

+++ src/ZenithEngine.Vulkan/VKCommandBuffer.cs
@@ -739,7 +739,7 @@ internal unsafe class VKCommandBuffer : CommandBuffer
     public override void DrawIndexed(uint indexCount,
                                      uint instanceCount,
                                      uint firstIndex = 0,
-                                     uint vertexOffset = 0,
+                                     int vertexOffset = 0,
                                      uint firstInstance = 0)
     {
         ValidatePipeline<VKGraphicsPipeline>(out _);
@@ -748,7 +748,7 @@ internal unsafe class VKCommandBuffer : CommandBuffer
                                   indexCount,
                                   instanceCount,
                                   firstIndex,
-                                  (int)vertexOffset,
+                                  vertexOffset,
                                   firstInstance);
     }
